# openapi.yaml
openapi: 3.1.0
info:
  title: Memoria API
  version: 1.0.0
  description: Private photo/video sharing API (Capsules + Albums). All times UTC in JSON.
servers:
  - url: http://localhost:8080
    description: Local docker-compose
  - url: https://api.memoria.app
    description: Production (placeholder)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
...
    userId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
paths:
...
  /v1/me/invite-link:
    post:
      tags: [friends]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: token }
...
  /v1/users/by-invite/{token}:
    get:
      tags: [friends]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Profile card }
        '404': { description: Not found }
  /v1/users/{username}:
    get:
      tags: [friends]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Profile + connection status }
  /v1/friends/requests:
    get:
      tags: [friends]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: incoming + outgoing }
    post:
      tags: [friends]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id: { type: string, format: uuid }
      responses:
        '201': { description: Request sent }
  /v1/friends/requests/{userId}/accept:
    post:
      tags: [friends]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '200': { description: Accepted }
  /v1/friends/requests/{userId}/decline:
    post:
      tags: [friends]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '200': { description: Declined }
  /v1/friends/requests/{userId}:
    delete:
      tags: [friends]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '204': { description: Cancelled }
  /v1/friends:
    get:
      tags: [friends]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: friends list }
  /v1/friends/{userId}:
    delete:
      tags: [friends]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '204': { description: Unfriended }
  /v1/blocks/{userId}:
    post:
      tags: [friends]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '200': { description: Blocked }
    delete:
      tags: [friends]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/userId'
      responses:
        '204': { description: Unblocked }
...